move context map to own file and remove add method
