+++ core/src/main/scala/zio/logging/ContextMap.scala
@@ -2,27 +2,10 @@ package zio.logging
 
 import zio._
 
-trait LoggingContext {
-  def loggingContext: LoggingContext.Service[Any]
-}
-
-object LoggingContext {
-  trait Service[-R] {
-    def get[V](key: ContextKey[V]): ZIO[R, Nothing, V]
-    def add[V](key: ContextKey[V], value: V): ZIO[R, Nothing, Unit]
-    def set[V](key: ContextKey[V], value: V): ZIO[R, Nothing, Unit]
-    def remove(key: ContextKey[_]): ZIO[R, Nothing, Unit]
-    def span[R1 <: R, E, A, V](key: ContextKey[V], value: V)(zio: ZIO[R1, E, A]): ZIO[R1, E, A]
-  }
-}
-
 final class ContextMap private (private val map: FiberRef[ContextMap.CMap]) extends LoggingContext.Service[Any] {
 
   def get[V](key: ContextKey[V]): UIO[V] =
     map.get.map(ContextMap.get(_)(key)._2)
-
-  def add[V](key: ContextKey[V], value: V): UIO[Unit] =
-    map.update(ContextMap.add(_)(key, value)).unit
 
   def set[V](key: ContextKey[V], value: V): UIO[Unit] =
     map.update(ContextMap.set(_)(key, value)).unit
